client-side_serviceacct/requirements.txt
# Python dependencies for Snowflake service account testing
snowflake-connector-python>=3.0.0
cryptography>=41.0.0

